BA Planner/v4/requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
pygetwindow>=0.0.9
pyautogui>=0.9.54
easyocr>=1.7.0
numpy>=1.24.0
